fix: align typed positioner state handling
Use overflow-safe calibrated step arithmetic, enforce the 0..60 stored-position range for typed MQTT commands, and apply step metadata and estimation consistently across command transports.

+++ software/nanoFramework/CubleyDiseqc.Managed/DiseqcPositionEstimate.cs
@@ -68,17 +68,17 @@ namespace Cubley.Diseqc
             int stepSize = direction == DiseqcMotorDirection.East
                 ? EastStepMicrodegrees
                 : WestStepMicrodegrees;
-            int signedDelta = direction == DiseqcMotorDirection.East
-                ? stepSize * steps
-                : -(stepSize * steps);
-            int target = EstimatedAngleMicrodegrees + signedDelta;
+            long signedDelta = direction == DiseqcMotorDirection.East
+                ? (long)stepSize * steps
+                : -((long)stepSize * steps);
+            long target = (long)EstimatedAngleMicrodegrees + signedDelta;
             if (target < -MaximumMagnitudeMicrodegrees || target > MaximumMagnitudeMicrodegrees)
             {
                 return;
             }
 
             HasPendingTarget = true;
-            PendingTargetMicrodegrees = target;
+            PendingTargetMicrodegrees = (int)target;
             PendingSource = "step";
         }
